+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 9.0.0.1 - (In Progress)
 
 *  Add `iswblank` and `iswxdigit` compatibility.
+*  Improve ctype compatibility.
 
 ## 9.0.0 - 2016-12-28